Rename 'settings.rs' to 'installed.rs'
