+++ src/commands/list.rs
@@ -1,6 +1,6 @@
 use prettytable::{cell, row, Cell, Row, Table};
 
-use crate::{selected::SelectedVersion, settings::InstalledToolchain, utils, Result};
+use crate::{installed::InstalledToolchain, selected::SelectedVersion, utils, Result};
 
 pub fn run(
     selected_version: &Option<SelectedVersion>,
