+++ src/installed.rs
