+++ src/shim.rs
@@ -3,7 +3,7 @@ use std::{env, ffi::OsString};
 use subprocess::{Exec, Redirection};
 
 use crate::{
-    dir_monitor::DirectoryMonitor, os::command_with_major_version, settings::InstalledToolchain,
+    dir_monitor::DirectoryMonitor, installed::InstalledToolchain, os::command_with_major_version,
     utils, Result, EXECUTABLE_NAME,
 };
