+++ src/utils.rs
@@ -18,8 +18,8 @@ use terminal_size::{terminal_size, Width};
 
 use crate::{
     constants::{home_env_variable, DEFAULT_DOT_DIR, EXECUTABLE_NAME, EXTRA_PACKAGES_FILENAME},
+    installed::InstalledToolchain,
     selected::SelectedVersion,
-    settings::InstalledToolchain,
     Result,
 };
